Made the calculator work in Rn space.

+++ dll/Database.cs
@@ -2,32 +2,32 @@ namespace lib;
 
 public static class Database
 {
-    public static Dictionary<string, Func<decimal, decimal, decimal>> BynaryStringToExpression = new()
+    public static Dictionary<string, Func<double, double, double>> BynaryStringToExpression = new()
     {
         {"sum" , (x,y)=>x+y},
         {"sub" , (x,y)=>x-y},
         {"mult" , (x,y)=>x*y},
         {"div" , (x,y)=>x/y},
-        {"pow" , (x,y)=>(decimal)Math.Pow((double)x,(double)y)}
+        {"pow" , (x,y)=>(double)Math.Pow((double)x,(double)y)}
         };
-    public static Dictionary<string, Func<decimal, decimal>> UnaryStringToExpression = new()
+    public static Dictionary<string, Func<double, double>> UnaryStringToExpression = new()
     {
-        {"arccos",(x)=>(decimal)Math.Acos((double)x)},
-        {"arcsin",(x)=>(decimal)Math.Asin((double)x)},
-        {"sinh",(x)=>(decimal)Math.Sinh((double)x)},
-        {"cosh",(x)=>(decimal)Math.Cosh((double)x)},
-        {"arcsinh",(x)=>(decimal)Math.Asinh((double)x)},
-        {"arccosh",(x)=>(decimal)Math.Acosh((double)x)},
-        { "sin",(x)=> (decimal)Math.Sin((double)x)},
-        { "cos",(x)=> (decimal)Math.Cos((double)x)},
-        { "tan",(x)=> (decimal)Math.Tan((double)x)},
-        { "cot",(x)=> (decimal)(1/Math.Tan((double)x))},
-        { "sec",(x)=> (decimal)(1/Math.Cos((double)x))},
-        { "csc",(x)=> (decimal)(1/Math.Sin((double)x))},
-        {"log2" , (x)=>(decimal)Math.Log((double)x,2)},
-        {"log10" , (x)=>(decimal)Math.Log10((double)x)},
-        {"log" , (x)=>(decimal)Math.Log10((double)x)},
-        {"ln" , (x)=>(decimal)Math.Log((double)x)}
+        {"arccos",(x)=>(double)Math.Acos((double)x)},
+        {"arcsin",(x)=>(double)Math.Asin((double)x)},
+        {"sinh",(x)=>(double)Math.Sinh((double)x)},
+        {"cosh",(x)=>(double)Math.Cosh((double)x)},
+        {"arcsinh",(x)=>(double)Math.Asinh((double)x)},
+        {"arccosh",(x)=>(double)Math.Acosh((double)x)},
+        { "sin",(x)=> (double)Math.Sin((double)x)},
+        { "cos",(x)=> (double)Math.Cos((double)x)},
+        { "tan",(x)=> (double)Math.Tan((double)x)},
+        { "cot",(x)=> (double)(1/Math.Tan((double)x))},
+        { "sec",(x)=> (double)(1/Math.Cos((double)x))},
+        { "csc",(x)=> (double)(1/Math.Sin((double)x))},
+        {"log2" , (x)=>(double)Math.Log((double)x,2)},
+        {"log10" , (x)=>(double)Math.Log10((double)x)},
+        {"log" , (x)=>(double)Math.Log10((double)x)},
+        {"ln" , (x)=>(double)Math.Log((double)x)}
     };
     public static Dictionary<string, string> binaryOperators = new()
     {
