+++ dll/Expression.cs
@@ -2,24 +2,24 @@ namespace lib;
 
 public abstract class Expression
 {
-    public abstract decimal Evaluate();
+    public abstract double Evaluate();
 }
 public class BinaryGeneric : Expression
 {
     Expression left;
     Expression right;
-    Func<decimal, decimal, decimal> func;
-    public BinaryGeneric(Expression left, Expression right, Func<decimal, decimal, decimal> func)
+    Func<double, double, double> func;
+    public BinaryGeneric(Expression left, Expression right, Func<double, double, double> func)
     {
         this.left = left;
         this.right = right;
         this.func = func;
     }
-    public override decimal Evaluate()
+    public override double Evaluate()
     {
         return Evaluate(left.Evaluate(), right.Evaluate());
     }
-    private decimal Evaluate(decimal left, decimal right)
+    private double Evaluate(double left, double right)
     {
         return func(left, right);
     }
@@ -27,48 +27,48 @@ public class BinaryGeneric : Expression
 public class UnaryGeneric : Expression
 {
     Expression expression;
-    Func<decimal, decimal> func;
-    public UnaryGeneric(Expression e, Func<decimal, decimal> func)
+    Func<double, double> func;
+    public UnaryGeneric(Expression e, Func<double, double> func)
     {
         this.expression = e;
         this.func = func;
     }
-    public override decimal Evaluate()
+    public override double Evaluate()
     {
         return Evaluate(expression.Evaluate());
     }
-    protected decimal Evaluate(decimal value)
+    protected double Evaluate(double value)
     {
         return this.func(value);
     }
 }
 public class Const : Expression
 {
-    readonly decimal value;
-    public Const(decimal value)
+    readonly double value;
+    public Const(double value)
     {
         this.value = value;
     }
 
-    public override decimal Evaluate()
+    public override double Evaluate()
     {
         return value;
     }
 }
 public class Variable : Expression
 {
-    private decimal x;
+    private double x;
     public Variable()
     {
         Evaluator.ChangeValue += WatchValue;
     }
 
-    public override decimal Evaluate()
+    public override double Evaluate()
     {
         return x;
     }
 
-    private void WatchValue(decimal x)
+    private void WatchValue(double x)
     {
         this.x = x;
     }
